Fix easy warnings in base
- Remove unsupported/unnecessary SuppressWarning annotations
- Remove reduntant type specifications (use diamond operator)
- Remove unused or duplicate imports
- Remove unnecessary casts (type is already correct type or can be
autoboxed)
- Remove unnecessary semi-colons (at end of class definitions, or just
repeated ones)
- Remove redundant super interfaces (interface that is already
inherited)
- Remove unused type parameters
- Remove declared checked exceptions that are never thrown
- Add missing @Override annotations

+++ modules/javafx.base/src/main/java/javafx/util/converter/LongStringConverter.java
@@ -63,6 +63,6 @@ public class LongStringConverter extends StringConverter<Long> {
             return "";
         }
 
-        return Long.toString(((Long)value).longValue());
+        return Long.toString(value.longValue());
     }
 }
